requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Session==0.3.1
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pyaml==18.11.0
PyYAML==3.13
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.20.1
smmap2==2.0.5
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
